+++ relay-spans/src/v1_to_v2.rs
@@ -55,10 +55,6 @@ pub fn span_v1_to_span_v2(span_v1: SpanV1) -> SpanV2 {
     attributes.insert("sentry.profile_id", profile_id.map_value(|v| v.to_string()));
     attributes.insert("sentry.platform", platform);
     attributes.insert("sentry.was_transaction", was_transaction);
-    attributes.insert(
-        "sentry._internal.performance_issues_spans",
-        performance_issues_spans,
-    );
 
     // Use same precedence as `backfill_data` for data bags:
     if let Some(measurements) = measurements.into_value() {
@@ -130,6 +126,7 @@ pub fn span_v1_to_span_v2(span_v1: SpanV1) -> SpanV2 {
         end_timestamp: timestamp,
         links: links.map_value(span_v1_links_to_span_v2_links),
         attributes: annotated_attributes,
+        performance_issues_spans,
         other: Default::default(), // cannot carry over because of schema mismatch
     }
 }
@@ -371,10 +368,6 @@ mod tests {
               "type": "string",
               "value": "{\"numbers\":[1,2,3]}"
             },
-            "sentry._internal.performance_issues_spans": {
-              "type": "boolean",
-              "value": true
-            },
             "sentry.client_sample_rate": {
               "type": "double",
               "value": 0.11
@@ -432,6 +425,7 @@ mod tests {
               "value": true
             }
           },
+          "_performance_issues_spans": true,
           "_meta": {
             "attributes": {
               "my.array": {